split account item components
