+++ src/components/user/voucher/account_management_tab.tsx
@@ -8,8 +8,6 @@ import {
   Loader2,
   AlertCircle,
   Pencil,
-  ChevronRight,
-  Trash2,
 } from "lucide-react";
 import { useParams } from "next/navigation";
 import { request } from "@/lib/utils/request";
@@ -21,274 +19,15 @@ import {
   IAccountingAccount,
   IAccountingAccountInput,
 } from "@/interfaces/accounting_account";
+import { ACCOUNT_TYPE_COLORS } from "@/constants/accounting_account";
 import SuccessNotification from "@/components/common/success_notification";
+import ConfirmModal from "@/components/common/confirm_modal";
+import {
+  CategorySubjectItem,
+  SubAccountItem,
+} from "@/components/user/voucher/account_item";
 
 // TODO: (20260703 - Julian) ============= 此元件還在施工中 =============
-
-// Info: (20260703 - Julian) 定義不同科目類別的顏色
-const ACCOUNT_TYPE_COLORS: Record<
-  string,
-  { bg: string; text: string; border: string; tab?: string }
-> = {
-  [AccountType.ASSET]: {
-    bg: "bg-emerald-50",
-    text: "text-emerald-700",
-    border: "border-emerald-200",
-    tab: "bg-emerald-500",
-  },
-  [AccountType.LIABILITY]: {
-    bg: "bg-rose-50",
-    text: "text-rose-700",
-    border: "border-rose-200",
-    tab: "bg-rose-500",
-  },
-  [AccountType.EQUITY]: {
-    bg: "bg-blue-50",
-    text: "text-blue-700",
-    border: "border-blue-200",
-    tab: "bg-blue-500",
-  },
-  [AccountType.REVENUE]: {
-    bg: "bg-amber-50",
-    text: "text-amber-700",
-    border: "border-amber-200",
-    tab: "bg-amber-500",
-  },
-  [AccountType.INCOME]: {
-    bg: "bg-teal-50",
-    text: "text-teal-700",
-    border: "border-teal-200",
-    tab: "bg-teal-500",
-  },
-  [AccountType.EXPENSE]: {
-    bg: "bg-lime-50",
-    text: "text-lime-700",
-    border: "border-lime-200",
-    tab: "bg-lime-500",
-  },
-  [AccountType.COST]: {
-    bg: "bg-pink-50",
-    text: "text-pink-700",
-    border: "border-pink-200",
-    tab: "bg-pink-500",
-  },
-  [AccountType.GAIN_OR_LOSS]: {
-    bg: "bg-violet-50",
-    text: "text-violet-700",
-    border: "border-violet-200",
-    tab: "bg-violet-500",
-  },
-  [AccountType.CASH_FLOW]: {
-    bg: "bg-cyan-50",
-    text: "text-cyan-700",
-    border: "border-cyan-200",
-    tab: "bg-cyan-500",
-  },
-  [AccountType.OTHER_COMPREHENSIVE_INCOME]: {
-    bg: "bg-fuchsia-50",
-    text: "text-fuchsia-700",
-    border: "border-fuchsia-200",
-    tab: "bg-fuchsia-500",
-  },
-  [AccountType.OTHER]: {
-    bg: "bg-slate-50",
-    text: "text-slate-700",
-    border: "border-slate-200",
-    tab: "bg-slate-500",
-  },
-};
-
-interface IAccountItemProps {
-  account: IAccountingAccount;
-  onAddChild: () => void;
-  onEdit: () => void;
-  onDelete?: () => void;
-  isSelected?: boolean;
-  onClick?: () => void;
-  hasChildren?: boolean;
-}
-
-/**
- * Info: (20260706 - Julian) 大類 (Level 1) 與 主科目 (Level 2) 元件
- */
-const CategorySubjectItem = ({
-  account,
-  onAddChild,
-  onEdit,
-  isSelected = false,
-  onClick = () => {},
-  hasChildren = false,
-}: IAccountItemProps) => {
-  const { t } = useTranslation();
-  const colors = ACCOUNT_TYPE_COLORS[account.type] || ACCOUNT_TYPE_COLORS.other;
-  const isLevel1 = account.level === 1;
-  const canClick = isLevel1 || hasChildren;
-
-  return (
-    <div
-      onClick={canClick ? onClick : undefined}
-      className={`group relative flex items-start gap-3 transition-all ${
-        isLevel1
-          ? "mt-6 mb-2 first:mt-0"
-          : `rounded-xl border px-4 py-4 shadow-sm ${
-              isSelected
-                ? "border-orange-200 bg-orange-50 ring-2 ring-orange-500 ring-offset-2"
-                : `border-slate-100 bg-white ${canClick ? "cursor-pointer hover:bg-gray-50" : "cursor-default"}`
-            } ${account.isCustom ? "border-dashed border-orange-300 bg-orange-100" : ""}`
-      }`}
-    >
-      {isLevel1 ? (
-        <div
-          className={`flex w-full items-center justify-between gap-2 rounded-lg border-b-2 border-slate-700 px-4 py-2.5 text-sm font-black tracking-widest text-white uppercase shadow-md ${colors.tab}`}
-        >
-          <p>{account.name}</p>
-          <p>({account.code})</p>
-        </div>
-      ) : (
-        <>
-          <div
-            className={`flex shrink-0 items-center justify-center rounded-lg border px-2 py-1 text-xs font-black ${colors.bg} ${colors.text} ${colors.border} mt-0.5`}
-          >
-            {account.code}
-          </div>
-          <div className="flex flex-1 items-start justify-between gap-1 overflow-hidden">
-            <div className="flex flex-1 flex-col overflow-hidden">
-              <div className="text-sm leading-relaxed font-bold wrap-break-word text-slate-700">
-                {account.name}
-              </div>
-              {account.isCustom && (
-                <span className="text-[10px] font-bold tracking-tighter text-orange-500 uppercase">
-                  {t("voucher.account.custom")}
-                </span>
-              )}
-            </div>
-            {hasChildren && (
-              <ChevronRight
-                size={16}
-                className={`mt-1 shrink-0 text-slate-300 transition-all ${
-                  isSelected ? "translate-x-1 text-orange-500" : ""
-                }`}
-              />
-            )}
-          </div>
-          {/* Hover Actions for Level 2 */}
-          <div className="absolute top-2 right-2 flex items-center gap-1 opacity-0 transition-opacity group-hover:opacity-100">
-            <button
-              onClick={(e) => {
-                e.stopPropagation();
-                onAddChild();
-              }}
-              className="flex size-7 items-center justify-center rounded-full bg-green-100 text-green-600 shadow-sm transition-colors hover:bg-green-200"
-              title={t("voucher.account.action.add_child")}
-            >
-              <Plus size={14} />
-            </button>
-            {account.isCustom && (
-              <button
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onEdit();
-                }}
-                className="flex size-7 items-center justify-center rounded-full bg-blue-50 text-blue-600 shadow-sm transition-colors hover:bg-blue-100"
-                title={t("voucher.account.action.edit")}
-              >
-                <Pencil size={12} />
-              </button>
-            )}
-          </div>
-        </>
-      )}
-    </div>
-  );
-};
-
-/**
- * Info: (20260706 - Julian) 子科目 (Level 3+) 元件
- */
-const SubAccountItem = ({
-  account,
-  onAddChild,
-  onEdit,
-  onDelete = () => {},
-}: IAccountItemProps) => {
-  const { t } = useTranslation();
-  const colors = ACCOUNT_TYPE_COLORS[account.type] || ACCOUNT_TYPE_COLORS.other;
-  const level = account.level;
-
-  return (
-    <div
-      className={`group relative flex items-center gap-3 rounded-xl border px-4 py-3 text-left shadow-sm transition-all md:px-6 ${
-        account.isCustom
-          ? "border-dashed border-orange-300 bg-orange-100"
-          : "border-slate-100 bg-white"
-      }`}
-      style={{ marginLeft: `${(level - 3) * 24}px` }}
-    >
-      <div
-        className={`flex shrink-0 items-center justify-center rounded-lg border px-3 py-1.5 text-sm font-bold md:text-base ${colors.bg} ${colors.text} ${colors.border}`}
-      >
-        {account.code}
-      </div>
-
-      <div className="flex flex-1 items-center justify-between gap-2 overflow-hidden">
-        <div className="flex flex-1 flex-col items-start gap-0.5 overflow-hidden">
-          <div className="text-base font-bold whitespace-normal text-slate-700">
-            {account.name}
-          </div>
-          <div className="flex items-center gap-2">
-            <div
-              className={`rounded-md border px-2 py-0.5 text-[10px] font-medium ${colors.bg} ${colors.text} ${colors.border}`}
-            >
-              {account.type}
-            </div>
-            {account.isCustom && (
-              <span className="text-[10px] font-bold text-orange-500 uppercase">
-                {t("voucher.account.custom")}
-              </span>
-            )}
-          </div>
-        </div>
-
-        <div className="flex items-center gap-1 opacity-0 transition-opacity group-hover:opacity-100">
-          <button
-            onClick={(e) => {
-              e.stopPropagation();
-              onAddChild();
-            }}
-            className="flex size-8 items-center justify-center rounded-full bg-green-100 text-green-600 transition-colors hover:bg-green-200"
-            title={t("voucher.account.action.add_child")}
-          >
-            <Plus size={16} />
-          </button>
-          {account.isCustom && (
-            <>
-              <button
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onEdit();
-                }}
-                className="flex size-8 items-center justify-center rounded-full bg-blue-50 text-blue-600 transition-colors hover:bg-blue-100"
-                title={t("voucher.account.action.edit")}
-              >
-                <Pencil size={14} />
-              </button>
-              <button
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onDelete?.();
-                }}
-                className="flex size-8 items-center justify-center rounded-full bg-red-50 text-red-600 transition-colors hover:bg-red-100"
-                title="刪除科目"
-              >
-                <Trash2 size={14} />
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const SUB_PAGE_SIZE = 10;
 
@@ -323,15 +62,21 @@ export default function AccountManagementTab() {
   const [formData, setFormData] =
     useState<IAccountingAccountInput>(emptyFormData);
 
+  // Info: (20260706 - Julian) Modal States
+  const [deleteId, setDeleteId] = useState<string | null>(null);
+  const [isConfirmModalOpen, setIsConfirmModalOpen] = useState<boolean>(false);
+
   useEffect(() => {
     const timer = setTimeout(() => {
       setDebouncedKeyword(keyword);
     }, 500);
 
-    // Info: (20260706 - Julian) 用關鍵字搜尋時，把主科目切換回全部，避免搜尋結果被限制
-    setActiveTab("all");
+    // Info: (20260706 - Julian) 關鍵字變更後，Tab 切回到 All
+    if (keyword && !debouncedKeyword) {
+      setActiveTab("all");
+    }
     return () => clearTimeout(timer);
-  }, [keyword]);
+  }, [keyword, debouncedKeyword]);
 
   const fetchAccounts = useCallback(async () => {
     setIsLoading(true);
@@ -368,35 +113,43 @@ export default function AccountManagementTab() {
     setSubPage(1);
   }, [selectedMainSubject]);
 
-  const handleAddChild = (parentAccount: IAccountingAccount) => {
+  const handleAddChild = useCallback((parentAccount: IAccountingAccount) => {
     setIsEditing(false);
     setEditId(null);
     setParentInfo(`[${parentAccount.code}] ${parentAccount.name}`);
     setFormData({ parentCode: parentAccount.code, name: "", code: "" });
     setErrorMessage(null);
-  };
+  }, []);
 
-  const handleEdit = (account: IAccountingAccount) => {
-    setIsEditing(true);
-    setEditId(account.id || null);
-    const parent = allAccounts.find((a) => a.code === account.parentCode);
-    setParentInfo(
-      parent ? `[${parent.code}] ${parent.name}` : account.parentCode,
-    );
-    setFormData({
-      parentCode: account.parentCode,
-      name: account.name,
-      code: account.code,
-    });
-    setErrorMessage(null);
-  };
+  const handleEdit = useCallback(
+    (account: IAccountingAccount) => {
+      setIsEditing(true);
+      setEditId(account.id || null);
+      const parent = allAccounts.find((a) => a.code === account.parentCode);
+      setParentInfo(
+        parent ? `[${parent.code}] ${parent.name}` : account.parentCode,
+      );
+      setFormData({
+        parentCode: account.parentCode,
+        name: account.name,
+        code: account.code,
+      });
+      setErrorMessage(null);
+    },
+    [allAccounts],
+  );
 
-  const handleDelete = async (id: string) => {
-    if (!window.confirm("確定要刪除此自訂科目嗎？")) return;
+  const handleDelete = useCallback((id: string) => {
+    setDeleteId(id);
+    setIsConfirmModalOpen(true);
+  }, []);
+
+  const confirmDelete = async () => {
+    if (!deleteId) return;
 
     try {
       const res = await request<IApiResponse<{ success: boolean }>>(
-        `/api/v1/user/account_book/${accountBookId}/accounting_account?id=${id}`,
+        `/api/v1/user/account_book/${accountBookId}/accounting_account?id=${deleteId}`,
         { method: "DELETE" },
       );
       if (res.success) {
@@ -408,6 +161,9 @@ export default function AccountManagementTab() {
     } catch (error) {
       console.error("Error deleting account", error);
       alert("刪除時發生錯誤");
+    } finally {
+      setIsConfirmModalOpen(false);
+      setDeleteId(null);
     }
   };
 
@@ -495,11 +251,17 @@ export default function AccountManagementTab() {
     return result;
   }, [selectedMainSubject, allAccounts]);
 
-  const totalSubPages = Math.ceil(rightList.length / SUB_PAGE_SIZE);
-  const paginatedSubAccounts = rightList.slice(
-    (subPage - 1) * SUB_PAGE_SIZE,
-    subPage * SUB_PAGE_SIZE,
-  );
+  const { totalSubPages, paginatedSubAccounts } = useMemo(() => {
+    const total = Math.ceil(rightList.length / SUB_PAGE_SIZE);
+    const paginated = rightList.slice(
+      (subPage - 1) * SUB_PAGE_SIZE,
+      subPage * SUB_PAGE_SIZE,
+    );
+    return {
+      totalSubPages: Math.max(1, total),
+      paginatedSubAccounts: paginated,
+    };
+  }, [rightList, subPage]);
 
   return (
     <div className="flex flex-col gap-6">
@@ -521,6 +283,7 @@ export default function AccountManagementTab() {
 
         <div className="flex flex-wrap items-center gap-1">
           <button
+            type="button"
             onClick={() => setActiveTab("all")}
             className={`${activeTab === "all" ? "bg-slate-800 text-white shadow-md" : "text-slate-500 hover:bg-gray-100"} rounded-lg px-4 py-2 text-xs font-bold transition-all md:text-sm`}
           >
@@ -528,6 +291,7 @@ export default function AccountManagementTab() {
           </button>
           {Object.entries(ACCOUNT_TYPE_COLORS).map(([key, value]) => (
             <button
+              type="button"
               key={key}
               onClick={() => setActiveTab(key as AccountType)}
               className={`${key === activeTab ? `${value.tab} text-white shadow-md` : `text-slate-500 hover:bg-slate-100`} rounded-lg px-4 py-2 text-xs font-bold transition-all md:text-sm`}
@@ -742,6 +506,16 @@ export default function AccountManagementTab() {
             : t("voucher.account.messages.create_success")
         }
         onClose={() => setShowSuccess(false)}
+      />
+
+      <ConfirmModal
+        isOpen={isConfirmModalOpen}
+        onClose={() => setIsConfirmModalOpen(false)}
+        title="刪除科目"
+        message="確定要刪除此自訂科目嗎？此動作無法復原。"
+        confirmText="確認刪除"
+        cancelText="取消"
+        onConfirm={confirmDelete}
       />
     </div>
   );
